# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: RDV.BI API
  version: 0.1.0
  summary: Appointment booking, queue, and institution administration API
  description: |
    OpenAPI contract for the current RDV.BI backend.

    Notes:
    - Public endpoints live under `/api/public`.
    - Authentication endpoints live under `/api/auth`.
    - Admin endpoints require a bearer JWT issued by `/api/auth/login`.
    - `POST /api/auth/bootstrap-admin` exists only for bootstrap and should be removed or disabled in production.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/admin/institutions/{institutionId}/counters:
    get:
      tags: [Admin Counters]
      summary: List counters for an institution
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
      responses:
        '200':
          description: Counters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Counter'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Admin Counters]
      summary: Create a counter
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CounterRequest'
      responses:
        '201':
          description: Created counter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Counter'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/counters/{id}:
    put:
      tags: [Admin Counters]
      summary: Update a counter
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CounterUpdateRequest'
      responses:
        '200':
          description: Updated counter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Counter'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Admin Counters]
      summary: Delete a counter
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/admin/users:
    get:
      tags: [Admin Users]
      summary: List users in the authenticated admin's institution
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Institution users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin Users]
      summary: Create a staff or admin user in the authenticated admin's institution
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/users/{id}:
    put:
      tags: [Admin Users]
      summary: Update a user in the authenticated admin's institution
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Admin Users]
      summary: Deactivate a user in the authenticated admin's institution
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidIdParam'
      responses:
        '204':
          description: User deactivated
        '404':
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InstitutionIdParam:
      name: institutionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to access the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example:
        error: Institution not found
...
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        password:
          type: string
          format: password
        active:
          type: boolean
      required: [email, full_name, role, password]
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        password:
          type: string
          format: password
        active:
          type: boolean
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
        institutionId:
          type: string
          format: uuid
        institutionName:
          type: string
      required: [id, email, fullName, role, active, institutionId, institutionName]
...
    CounterRequest:
      type: object
      properties:
        number:
          type: string
          maxLength: 50
        staff_name:
          type: string
          maxLength: 255
        is_active:
          type: boolean
      required: [number]
    CounterUpdateRequest:
      type: object
      properties:
        number:
          type: string
        staff_name:
          type: string
        is_active:
          type: boolean
        status:
          $ref: '#/components/schemas/CounterStatus'
        current_ticket:
          type: string
...
    Counter:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        staffName:
          type: string
        status:
          $ref: '#/components/schemas/CounterStatus'
        active:
          type: boolean
        currentTicket:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, number, status, active]
...
    CounterStatus:
      type: string
      enum: [AVAILABLE, BUSY, BREAK, CLOSED]
...
    UserRole:
      type: string
      enum: [ADMIN, STAFF]